ringbuffer_process (ringbuffer v1.2.0+build.18.refca920a0)

Process to own the created ets table for the ring buffer.

Functions

code_change(OldVsn, State, Extra)

handle_call(_, From, State)

handle_cast(Msg, State)

handle_info(Info, State)

init(_)

process_pid(Name)

-spec process_pid(Name :: atom()) -> {ok, pid()} | {error, badarg}.

Find the process pid for the process owning the ets table. If the buffer does not exist then an error is returned.

read(Name)

-spec read(Name :: atom()) -> {ok, {Skipped :: non_neg_integer(), Payload :: term()}} | {error, empty}.

Read a payload from the buffer. Skip entries that are deleted by the writers.

There are a coupe of race conditions we need to take care of:

1. Writer incremented -> Reader tries entry --> Writer writes entry In this case the w-value read by the reader is Size smaller than that of the writer. In this case we wait till the writer finishes writing.

2. Two or more readers arrive at the same time, incrementing the reader value. One of the readers will have the correct value and can fetch the next entry. In this case the reader can move past the writer if there are not enough entries in the buffer for all readers.

For the case where multiple readers are racing past the writer can not be solved without synchronization, we let the ringbuffer process handle the reader increment.

start_link(Name, Size)

-spec start_link(atom(), pos_integer()) -> {ok, pid()} | {error, term()}.

Create a new process managing a ringbuffer.

terminate(Reason, State)

write(Name, Payload)

-spec write(Name :: atom(), Payload :: term()) -> ok.

Write a payload the buffer. If the readers can't keep up then older entries are deleted.